+++ src/Helper.php
@@ -2,6 +2,8 @@
 
 namespace Drupal\bootstrap_italia;
 
+use Drupal\Core\Config\ImmutableConfig;
+
 /**
  * Helper class for bootstrap_italia theme.
  */
@@ -148,23 +150,25 @@ class Helper {
   }
 
   /**
-   * Returns available themes variants.
+   * Returns available theme libraries.
    *
    * @return array
-   *   Array with theme variants options.
+   *   Array with theme libraries options.
    */
-  public static function getThemeVariants(): array {
+  public static function getThemeLibraries(): array {
     $theme = \Drupal::service('theme.manager')->getActiveTheme();
     $assets_css_path = $theme->getPath() . '/assets/css';
     $assets_js_path = $theme->getPath() . '/assets/js';
 
-    $default_option = [
-      '' => t('Use @theme.libraries.yml (recommended)', ['@theme' => $theme->getName()]),
+    $static_option = [
+      '' => t('Use @theme.libraries.yml', ['@theme' => $theme->getName()]),
+      'vanilla' => t('Vanilla libraries'),
+      'cdn' => t('CDN libraries'),
     ];
 
     // Check if directories css and js exists.
     if (!is_dir($assets_css_path) || !is_dir($assets_js_path)) {
-      return $default_option;
+      return $static_option;
     }
 
     // Config extensions parameters.
@@ -176,54 +180,130 @@ class Helper {
     $css = \Drupal::service('file_system')->scanDirectory($assets_css_path, "/{$extensions}$/");
     $js = \Drupal::service('file_system')->scanDirectory($assets_js_path, "/{$extensions}$/");
 
-    // Make a variants return array.
-    $variants = [];
-    $variants += $default_option;
+    // Make a libraries return array.
+    $libraries = [];
+    $libraries += $static_option;
 
-    // Loop for css file.
+    // Loop for css files.
     foreach ($css as $fileC) {
+      // Loop for js files.
       foreach ($js as $fileJ) {
         // If name css and js match.
         if ($fileC->name == $fileJ->name) {
           // Add variant.
-          $variants[$fileC->name] = $fileC->name . '.[css/js]';
+          $libraries[$fileC->name] = 'Webpack: ' . $fileC->name . '.[css/js]';
         }
       }
     }
-    return $variants;
+    return $libraries;
   }
 
   /**
-   * Set themes variants.
+   * Get config from Drupal Service.
+   *
+   * @return \Drupal\Core\Config\ImmutableConfig
+   *   Current default theme config.
+   */
+  public static function getThemeSettings(): ImmutableConfig {
+    // Get default theme.
+    $theme = \Drupal::service('theme.manager')->getActiveTheme()->getName();
+    // Get config default theme.
+    return \Drupal::config($theme . '.settings');
+  }
+
+  /**
+   * Set themes libraries.
    *
    * @return array
    *   Libraries array.
    */
-  public static function setThemeVariants(): array {
-    // Get default theme.
-    $theme = \Drupal::service('theme.manager')->getActiveTheme()->getName();
-    // Get config default theme.
-    $config = \Drupal::config($theme . '.settings');
-    // Get theme_variant default theme.
-    $theme_variant = $config->getOriginal('theme_variant');
+  public static function setThemeLibraries(): array {
+    // Get libraries_type from Drupal config.
+    $libraries_type = self::getThemeSettings()->getOriginal('libraries_type');
 
-    $libraries = [];
-    if (!empty($theme_variant) && $theme_variant != '') {
-      $libraries['global'] = [
+    if (!empty($libraries_type) && $libraries_type == 'vanilla') {
+      $libraries['libraries-ui'] = self::getThemeLibrariesVanilla();
+    }
+    elseif (!empty($libraries_type) && $libraries_type == 'cdn') {
+      $libraries['libraries-ui'] = self::getThemeLibrariesCDN();
+    }
+    elseif (!empty($libraries_type) && $libraries_type != '') {
+      $libraries['libraries-ui'] = [
         'css' => [
           'theme' => [
-            'assets/css/' . $theme_variant . '.css' => [],
+            'assets/css/' . $libraries_type . '.css' => [],
           ],
         ],
         'js' => [
-          'assets/js/' . $theme_variant . '.js' => [],
+          'assets/js/' . $libraries_type . '.js' => [],
         ],
         'dependencies' => [
           'core/drupal',
         ],
       ];
     }
+    else {
+      $libraries = [];
+    }
+//    dump($libraries);
     return $libraries;
+  }
+
+  /**
+   * Returns vanilla libraries.
+   *
+   * @return array
+   *   Array with vanilla libraries
+   */
+  public static function getThemeLibrariesVanilla(): array {
+    $js = self::getThemeSettings()->getOriginal('libraries_bundle')
+      ? 'assets/js/bootstrap-italia.bundle.min.js'
+      : 'assets/js/bootstrap-italia.min.js';
+
+    return [
+      'css' => [
+        'theme' => [
+          'assets/css/bootstrap-italia.min.css' => ['minified' => TRUE]
+        ],
+      ],
+      'js' => [
+        $js => ['minified' => TRUE]
+      ],
+      'dependencies' => [
+        'core/drupal',
+      ],
+    ];
+  }
+
+  /**
+   * Returns CDN libraries.
+   *
+   * @return array
+   *   Array with CDN libraries
+   */
+  public static function getThemeLibrariesCDN(): array {
+    $css = self::getThemeSettings()->getOriginal('libraries_cdn_css');
+    $js = self::getThemeSettings()->getOriginal('libraries_cdn_js');
+    $min = self::getThemeSettings()->getOriginal('libraries_cdn_minified');
+    return [
+      'css' => [
+        'theme' => [
+          $css => [
+            'type' => 'external',
+            'minified' => $min
+          ]
+        ],
+      ],
+      'js' => [
+        $js => [
+          'type' => 'external',
+          'minified' => $min
+        ]
+      ],
+      'dependencies' => [
+        'core/drupal',
+      ],
+    ];
   }
 
 }